% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taurus.R
\name{aplicar_regra_1}
\alias{aplicar_regra_1}
\title{Aplica a Regra 1 para armas da marca Taurus}
\usage{
aplicar_regra_1(tab, calibre)
}
\arguments{
\item{tab}{Data frame contendo os dados das armas.}

\item{calibre}{Calibre da arma a ser considerado na regra.}
}
\value{
Data frame com as colunas \code{id_bo}, \code{id_arma}, \code{arma_ano_fabricacao} e \code{padrao_taurus}. A coluna \code{padrao_taurus} indica se a arma segue a Regra 1.
}
\description{
Esta função aplica a Regra 1 para um conjunto de dados de armas da marca Taurus.
A função filtra as armas que possuem número de série disponível, converte o número
de série para formato numérico, ajusta o calibre da arma e realiza um join com uma
tabela de referência para determinar o ano de fabricação da arma.
}
