% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taurus.R
\name{aplicar_regra_2_1}
\alias{aplicar_regra_2_1}
\title{Aplica a Regra 2.1 para Filtragem e Transformação de Dados de Armas}
\usage{
aplicar_regra_2_1(tab)
}
\arguments{
\item{tab}{Data frame contendo os dados das armas.}
}
\value{
Data frame transformado após a aplicação da Regra 2.1.
}
\description{
Esta função aplica a Regra 2.1 para filtrar e transformar dados de armas da
marca Taurus.
}
\details{
A função realiza as seguintes operações:
\enumerate{
\item Filtra as armas que possuem número de série disponível, número de série
formatado com 8 caracteres, começando com três letras maiúsculas, marca
final "taurus" e tipo "pistola".
\item Extrai a segunda e terceira letras do número de série formatado.
\item Realiza um join com a tabela depara da Regra 2.
\item Remove duplicatas com base nos campos \code{id_bo}, \code{id_arma} e
\code{arma_ano_fabricacao}.
\item Converte o ano de fabricação da arma para caractere e adiciona a coluna
\code{padrao_taurus} com o valor "Regra 2.1" se o ano de fabricação não for NA.
}
}
