% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taurus.R
\name{aplicar_regra_2_2}
\alias{aplicar_regra_2_2}
\title{Aplica a Regra 2.2 para Filtragem e Mutação de Dados de Armas}
\usage{
aplicar_regra_2_2(tab)
}
\arguments{
\item{tab}{Data frame contendo os dados das armas.}
}
\value{
Data frame modificado após a aplicação da Regra 2.2.
}
\description{
Esta função aplica a Regra 2.2 a uma tabela de dados de armas, filtrando e
modificando os dados conforme especificado.
}
\details{
A função realiza as seguintes
operações:
\enumerate{
\item Filtra as armas que possuem número de série disponível, número de série
formatado com 7 ou 8 caracteres, que começam com duas letras seguidas de um
número,  marca "taurus" e tipo "revolver".
\item Adiciona colunas com a primeira e segunda letra do número de série.
\item Realiza um join com a tabela depara da regra 2.
\item Remove duplicatas com base em \code{id_bo}, \code{id_arma} e \code{arma_ano_fabricacao}.
\item Converte o ano de fabricação para caractere e adiciona a coluna \code{padrao_taurus}
indicando "Regra 2.2" quando o ano de fabricação não é NA.
}
}
