% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taurus.R
\name{aplicar_regra_3}
\alias{aplicar_regra_3}
\title{Aplica a Regra 3 para a Tabela de Armas}
\usage{
aplicar_regra_3(tab)
}
\arguments{
\item{tab}{Um data frame contendo os dados das armas.}
}
\value{
Um data frame com as colunas \code{id_bo}, \code{id_arma}, \code{arma_ano_fabricacao} e \code{padrao_taurus} após a aplicação da Regra 3.
}
\description{
Esta função aplica a Regra 3 para uma tabela de armas, filtrando e transformando os dados conforme necessário.
}
\details{
A função \code{aplicar_regra_3} realiza as seguintes operações:
\enumerate{
\item Filtra as armas que possuem número de série formatado, com exatamente 9 caracteres, começando com três letras maiúsculas, e cuja marca final é "taurus".
\item Extrai as três primeiras letras do número de série formatado.
\item Realiza um join com a tabela depara da Regra 3, utilizando as três primeiras letras do número de série.
\item Remove duplicatas, mantendo apenas as colunas \code{id_bo}, \code{id_arma} e \code{arma_ano_fabricacao}.
\item Converte o ano de fabricação para caractere e adiciona a coluna \code{padrao_taurus} com o valor "Regra 3" para registros com ano de fabricação não nulo.
}
}
