% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodo.R
\name{categorizar_periodo}
\alias{categorizar_periodo}
\title{Categoriza o período do dia com base em um horário fornecido}
\usage{
categorizar_periodo(x)
}
\arguments{
\item{x}{Um vetor de caracteres representando horários no formato "HH:MM"}
}
\value{
Um vetor de caracteres com a categoria do período do dia correspondente:
\describe{
\item{Madrugada}{Para horários entre "00:00" e "04:00"}
\item{Início Manhã}{Para horários até "08:00"}
\item{Manhã}{Para horários até "12:00"}
\item{Início Tarde}{Para horários até "16:00"}
\item{Tarde}{Para horários até "20:00"}
\item{Noite}{Para horários entre "20:00" e "23:59"}
}
}
\description{
Categoriza o período do dia com base em um horário fornecido
}
\examples{
categorizar_periodo("03:00") # Retorna "Madrugada"
categorizar_periodo("09:00") # Retorna "Início Manhã"
categorizar_periodo("15:00") # Retorna "Início Tarde"

}
