% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depara.R
\name{depara_tipo}
\alias{depara_tipo}
\alias{depara_calibre}
\alias{depara_calibre_policial}
\alias{depara_marca}
\alias{depara_crime}
\title{Funções de-para}
\usage{
depara_tipo(tab, nome_coluna = "arma_tipo")

depara_calibre(tab, nome_coluna = "calibre")

depara_calibre_policial(tab, base)

depara_marca(tab, nome_coluna = "marca")

depara_crime(tab, nome_coluna)
}
\arguments{
\item{tab}{Data frame. A tabela que será unida com a tabela de "de-para".}

\item{nome_coluna}{Character. O nome da coluna na tabela \code{tab} que contém a chave para o de-para.}

\item{base}{Character. A base de dados que será utilizada para o de-para. Pode ser "rj_complementar" ou "rj_principal".}
}
\value{
Data frame. A tabela original \code{tab} com as novas colunas provinientes do de-para.
}
\description{
Estas funções realizam a junção de uma tabela com uma tabela de "de-para".
}
