% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolocalizacao.R
\name{geolocalizar}
\alias{geolocalizar}
\title{Geolocalizar endereços}
\usage{
geolocalizar(rua = "", cidade = "", estado = "", cep = "", pais = "")
}
\arguments{
\item{rua}{Vetor de caracteres com os nomes das ruas.}

\item{cidade}{Vetor de caracteres com os nomes das cidades.}

\item{estado}{Vetor de caracteres com os nomes dos estados.}

\item{cep}{Vetor de caracteres com os códigos postais (CEP).}

\item{pais}{Vetor de caracteres com os nomes dos países.}
}
\value{
Um data frame com as colunas \code{lat} e \code{long} contendo as coordenadas geográficas dos endereços fornecidos.
}
\description{
Esta função realiza a geolocalização de endereços fornecidos, retornando as coordenadas de latitude e longitude.
}
\examples{
\dontrun{
ruas <- c("Av. Paulista", "Praça da Sé")
cidades <- c("São Paulo", "São Paulo")
estados <- c("SP", "SP")
ceps <- c("01311-000", "01001-000")
paises <- c("Brasil", "Brasil")

geolocalizar(rua = ruas, cidade = cidades, estado = estados, cep = ceps, pais = paises)
}
}
