% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolocalizacao.R
\name{geolocalizar_endereco}
\alias{geolocalizar_endereco}
\title{Geolocalizar Endereço}
\usage{
geolocalizar_endereco(endereco)
}
\arguments{
\item{endereco}{Vetor de caracteres contendo os endereços a serem geolocalizados.}
}
\value{
Data frame com as colunas de latitude e longitude dos endereços geolocalizados.
}
\description{
Esta função recebe um vetor de endereços e realiza a geolocalização de cada um deles,
salvando os resultados em um arquivo CSV e retornando as coordenadas (latitude e longitude).
}
\examples{
\dontrun{
enderecos <- c("Rua Exemplo 123, Cidade, Estado", "Avenida Teste 456, Cidade, Estado")
coordenadas <- geolocalizar_endereco(enderecos)
print(coordenadas)
}

}
