% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_calibre.R
\name{gerar_arma_calibre_final}
\alias{gerar_arma_calibre_final}
\title{Gerar Calibre Final da Arma}
\usage{
gerar_arma_calibre_final(tab)
}
\arguments{
\item{tab}{Um data frame contendo as colunas \code{calibre_formatado_final} e
\code{tipo_formatado} (colunas criadas pela função \code{\link[=depara_calibre]{depara_calibre()}} e
\code{\link[=depara_tipo]{depara_tipo()}} respectivamente).}
}
\value{
Um data frame com uma coluna adicional \code{arma_calibre_final} que contém o calibre final da arma.
}
\description{
Esta função processa uma tabela de armas para gerar a coluna \code{arma_calibre_final}
a partir de regras relacionadas ao \code{calibre_formatado_final} e ao \code{tipo_formatado}.
}
\details{
As condições para determinar o calibre final são as seguintes:
\itemize{
\item Se \code{calibre_formatado_final} for ".32" e \code{tipo_formatado} for "espingarda", o calibre final é "32 gauge".
\item Se \code{calibre_formatado_final} for ".32" e \code{tipo_formatado} for "revolver" ou "garrucha", o calibre final é ".32 S&W long".
\item Se \code{calibre_formatado_final} for ".32" e \code{tipo_formatado} for "pistola", "carabina" ou "submetralhadora", o calibre final é ".32 acp".
\item Se \code{calibre_formatado_final} terminar com "32", o calibre final é ".32 S&W long".
\item Caso contrário, o calibre final é o valor de \code{calibre_formatado_final}.
}

A coluna \code{arma_calibre_final} resultante é então convertida para maiúsculas.
}
