% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_marca.R
\name{gerar_arma_marca_final}
\alias{gerar_arma_marca_final}
\title{Gerar Marca Final da Arma}
\usage{
gerar_arma_marca_final(tab)
}
\arguments{
\item{tab}{Um data frame contendo a coluna \code{arma_marca} que possui as informações da marca da arma (coluna criada pela função \code{\link[=depara_marca]{depara_marca()}}).}
}
\value{
Um data frame com uma coluna adicional \code{arma_marca_final} que contém os nomes padronizados das marcas de armas.
}
\description{
Esta função processa uma tabela fornecida para gerar uma coluna final de marca de arma (\code{arma_marca_final}) com base nos valores da coluna \code{arma_marca}.
}
\details{
A função aplica regras para a criação da coluna final de calibre,
que possivelmente sobrescrevem os padrões definidos pela função \code{\link[=depara_marca]{depara_marca()}}.
Ela faz isso definindo uma marca específica à arma quando um certo padrão é detectado na marca original. Os seguintes mapeamentos são aplicados:
\itemize{
\item "taur" -> "taurus"
\item "ross" -> "rossi"
\item "glo[ck]" -> "glock"
\item "smith" -> "s&w"
}
Se nenhum desses padrões for detectado, a função atribui o valor da coluna \code{marca_arma_v2} (convertido para minúsculas) à \code{arma_marca_final}.
}
