% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_tipo.R
\name{gerar_arma_tipo_final}
\alias{gerar_arma_tipo_final}
\title{Gera a tabela final de tipos de armas}
\usage{
gerar_arma_tipo_final(tab)
}
\arguments{
\item{tab}{Um data frame contendo as colunas \code{tipo_formatado}
(gerada pela função \code{\link[=depara_tipo]{depara_tipo()}}) e \code{tipo_arma_calibre} (gerada pela
função \code{\link{depara_calibre}}).}
}
\value{
Data frame com a tabela final de tipos de armas, com colunas renomeadas e formatadas.
}
\description{
Esta função recebe uma tabela e realiza transformações nas colunas para gerar a tabela final de tipos de armas.
}
