% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_arma.R
\name{gerar_flag_arma}
\alias{gerar_flag_arma}
\title{Gerar Flag de Arma}
\usage{
gerar_flag_arma(tab)
}
\arguments{
\item{tab}{Um data frame contendo a coluna \code{compatibilidade_tipo} e \code{flag_arma}.}
}
\value{
Um data frame com a coluna \code{flag_arma} atualizada.
}
\description{
Esta função gera uma flag de arma em uma tabela fornecida. A flag é definida como lógica e ajustada com base na compatibilidade do tipo de arma.
}
