% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_arma.R
\name{gerar_flag_arma_artesanal}
\alias{gerar_flag_arma_artesanal}
\title{Gera flag para arma artesanal}
\usage{
gerar_flag_arma_artesanal(tab)
}
\arguments{
\item{tab}{Um data frame contendo uma coluna \code{tipo_formatado} com a descrição do tipo de arma (gerada pela função \code{\link[=depara_tipo]{depara_tipo()}}).}
}
\value{
Um data frame com uma nova coluna \code{flag_arma_artesanal}, que é TRUE se o tipo de arma for artesanal e FALSE caso contrário.
}
\description{
Esta função adiciona uma coluna \code{flag_arma_artesanal} à tabela fornecida,
indicando se o tipo de arma é artesanal.
}
