% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_arma_policial.R
\name{gerar_flag_arma_policia_prop}
\alias{gerar_flag_arma_policia_prop}
\title{Gerar Flag de Arma de Propriedade Policial}
\usage{
gerar_flag_arma_policia_prop(tab, base)
}
\arguments{
\item{tab}{Data frame contendo os dados a serem processados.}

\item{base}{String indicando a base de dados a ser utilizada. Pode ser "rj_complementar" ou "sp".}
}
\value{
Data frame com as colunas adicionais \code{flag_arma_policia_prop},
\code{flag_arma_policia_desd} (apenas para a base "sp") e \code{flag_arma_policia_prop_indisp} (apenas para a base "sp").
}
\description{
Esta função gera uma flag indicando se a arma pertence à polícia ou não,
com base em diferentes critérios dependendo da base de dados fornecida.
}
\details{
Para a base "rj_complementar", a função verifica se a coluna \code{patrimoniada} contém
valores que indicam que a arma pertence a alguma força policial ou militar.

Para a base "sp", a função utiliza expressões regulares para verificar se a coluna
\code{arma_proprietario_nome} contém termos que indicam propriedade policial, além de
verificar se a coluna \code{desdobramento} indica intervenção policial.
}
