% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_arma_policial.R
\name{gerar_flag_arma_policial}
\alias{gerar_flag_arma_policial}
\title{Gerar Flag de Arma Policial}
\usage{
gerar_flag_arma_policial(tab, base)
}
\arguments{
\item{tab}{Data frame contendo os dados a serem processados.}

\item{base}{String indicando a base de dados a ser utilizada. Pode ser "rj", "rj_complementar" ou "sp".}
}
\value{
Data frame com a flag de arma policial adicionada.
}
\description{
Esta função gera uma flag indicando se uma arma é policial com base em diferentes critérios dependendo da base de dados fornecida.
}
\details{
\itemize{
\item Para a base "rj", a flag é gerada com base na rubrica formatada e no calibre policial.
\item Para a base "rj_complementar", a flag é gerada com base na disponibilidade
de número de série, na propriedade da arma pela polícia e no calibre policial.
\item Para a base "sp", a flag é gerada com base na disponibilidade de número de
série, na propriedade da arma pela polícia e no calibre policial.
}
}
