% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_arma_policial.R
\name{gerar_flag_mdoip}
\alias{gerar_flag_mdoip}
\title{Gerar Flag MDOIP}
\usage{
gerar_flag_mdoip(tab)
}
\arguments{
\item{tab}{Um data frame contendo os dados a serem processados.}
}
\value{
Um data frame com uma nova coluna \code{flag_mdoip} indicando se o registro é um MDOIP.
}
\description{
Esta função gera uma flag para identificar registros de MDOIP
(Morte Decorrente de Intervenção Policial) em uma tabela de dados.
}
