% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_tipo.R
\name{gerar_flag_tipo_arma_incompativel}
\alias{gerar_flag_tipo_arma_incompativel}
\title{Gerar Flag de Tipo de Arma Incompatível}
\usage{
gerar_flag_tipo_arma_incompativel(tab)
}
\arguments{
\item{tab}{Um data frame contendo as colunas \code{tipo_formatado}
(gerada pela função \code{\link[=depara_tipo]{depara_tipo()}}) e \code{tipo_arma_calibre} (gerada pela
função \code{\link{depara_calibre}}).}
}
\value{
Um data frame com uma nova coluna \code{flag_tipo_arma_incompativel_calibre} que indica se o tipo de arma é incompatível com o calibre.
}
\description{
Esta função gera uma flag indicando se o tipo de arma é incompatível com o calibre informado.
}
