% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_arma.R
\name{gerar_id_arma}
\alias{gerar_id_arma}
\title{Gerar Identificador de Arma}
\usage{
gerar_id_arma(tab, base)
}
\arguments{
\item{tab}{Um data frame contendo os dados das armas.}

\item{base}{Uma string indicando a base de dados. Pode ser "rj", "rj_complementar" ou "sp".}
}
\value{
Um data frame com uma nova coluna \code{id_arma} contendo os identificadores únicos.
}
\description{
Esta função gera um identificador único para cada arma em uma tabela,
baseado em diferentes colunas dependendo da base de dados fornecida.
}
