% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_bo.R
\name{gerar_id_bo}
\alias{gerar_id_bo}
\title{Gerar ID de Boletim de Ocorrência (BO)}
\usage{
gerar_id_bo(tab, base)
}
\arguments{
\item{tab}{Um data frame contendo os dados dos boletins de ocorrência.}

\item{base}{Uma string que especifica a base de dados. Pode ser "rj_complementar" ou "rj".}
}
\value{
Um data frame com uma nova coluna \code{id_bo} contendo os identificadores
únicos para os boletins de ocorrência, se a base for "rj_complementar" ou "rj".
Caso contrário, retorna o data frame original sem modificações.
}
\description{
Esta função gera um identificador único para boletins de ocorrência (BO)
com base em diferentes colunas, dependendo da base de dados especificada.
}
