% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_numero_serie.R
\name{gerar_numero_serie_formatado}
\alias{gerar_numero_serie_formatado}
\title{Gerar Número de Série Formatado}
\usage{
gerar_numero_serie_formatado(tab)
}
\arguments{
\item{tab}{Um data frame contendo a coluna \code{sn_disponivel} e \code{arma_numero_serie}.}
}
\value{
Um data frame com uma nova coluna \code{arma_numero_serie_formatado} onde:
\itemize{
\item Se \code{sn_disponivel} for "Sim", o número de série será convertido para letras maiúsculas e espaços serão removidos.
\item Caso contrário, a coluna será preenchida com uma string vazia.
}
}
\description{
Esta função formata o número de série das armas em uma tabela.
}
