% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubrica.R
\name{gerar_rubrica_formatada}
\alias{gerar_rubrica_formatada}
\title{Gerar Rubrica Formatada}
\usage{
gerar_rubrica_formatada(tab, base)
}
\arguments{
\item{tab}{DataFrame. A tabela que será formatada.}

\item{base}{Character. A base de dados que será utilizada para formatar a tabela. Pode ser "rj" ou "rj_complementar".}
}
\value{
DataFrame. A tabela formatada de acordo com a base especificada.
}
\description{
Esta função formata a tabela de acordo com a base especificada.
}
