% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_numero_serie.R
\name{gerar_sn_disponivel}
\alias{gerar_sn_disponivel}
\title{Gerar Disponibilidade do Número de Série da Arma}
\usage{
gerar_sn_disponivel(tab)
}
\arguments{
\item{tab}{Um data frame contendo uma coluna \code{arma_numero_serie} com os números de série das armas.}
}
\value{
Um data frame com uma nova coluna \code{sn_disponivel} que classifica a disponibilidade do número de série da arma em várias categorias, como "não aparente", "removido", "raspado", "puncionamento", "regravado/adulterado", "desgastado", "só pontuações", "só texto" e "Sim".
}
\description{
Esta função recebe uma tabela e adiciona uma coluna \code{sn_disponivel} que indica a disponibilidade do número de série da arma.
}
\examples{
\dontrun{
tab <- data.frame(arma_numero_serie = c("1234", "expu123", "raspa567", "abcd", "1234abcd"))
gerar_sn_disponivel(tab)
}
}
