% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depara.R
\name{ler_depara}
\alias{ler_depara}
\title{Ler De/Para}
\usage{
ler_depara(aba)
}
\arguments{
\item{aba}{Nome da aba a ser lida no arquivo Excel.}
}
\value{
Um data frame contendo os dados da aba especificada.
}
\description{
Esta função lê uma aba específica de um arquivo Excel que contém
as tabelas de de/para.
}
\examples{
\dontrun{
  dados <- ler_depara("calibre")
}
}
