% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolocalizacao.R
\name{montar_endereco}
\alias{montar_endereco}
\title{Montar Endereço RJ}
\usage{
montar_endereco(tab, base)
}
\arguments{
\item{tab}{Um data frame contendo as colunas \code{bairro_fato} e \code{municipio_fato}.}

\item{base}{String indicando a base de dados a ser utilizada. Pode ser "rj", "rj_complementar" ou "sp".}
}
\value{
Um data frame com colunas adicionais: \code{estado}, \code{pais} e \code{endereco}.
}
\description{
Esta função recebe uma tabela e monta uma string de endereço.
}
